pyproject.toml
[project]
name = "md2html"
version = "0.1.0"
description = "Markdown to HTML converter with theme and custom containers"
authors = [{name = "GitHub Copilot"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "Jinja2>=3.1.0",
    "PyYAML>=6.0",
    "Pygments>=2.16",
    "watchdog>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
md2html = "md2html.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
